refactor(designer): extract the compartment grid's editing model
useCompartmentGrid holds what the grid IS and how pointer input changes it
(cell ownership, drag-to-merge/split, the 3D ghost preview, divider hit
targets). CompartmentGridView draws it and takes its sizing from the caller.
CompartmentEditor keeps only the sidebar's own chrome: grid steppers, the
by-size solver, wall thickness and the divider subsections.

Shared rather than duplicated because the Bento workspace has to draw the same
grid: a merge rule that differed between the two surfaces would be a silent
correctness bug, not a styling difference.

No behaviour change. The 75 existing CompartmentEditor tests pass untouched.
