+++ src/features/bin-designer/components/CompartmentEditor/CompartmentEditor.tsx
@@ -1,5 +1,5 @@
 /**
- * Visual compartment grid editor.
+ * Visual compartment grid editor (sidebar surface).
  *
  * Displays a top-down 2D view of the bin interior divided into a user-defined
  * grid. Users can:
@@ -11,46 +11,33 @@
  * form one rectangular compartment. Divider walls are automatically derived
  * from boundaries between cells with different IDs.
  *
- * Sub-components live in `CompartmentEditorParts.tsx` (`GridCell` and
- * `GhostPreview`). Preview color sync is centralized in the shared
- * `usePreviewColor` hook.
+ * The editing model lives in `useCompartmentGrid` and the grid itself in
+ * `CompartmentGridView`, both shared with the full-size Bento workspace. What
+ * remains here is the sidebar's own chrome: the grid steppers, the by-size
+ * solver, wall thickness and the divider subsections.
+ *
+ * Cell and ghost sub-components live in `CompartmentEditorParts.tsx`.
  */
 
-import { useCallback, useEffect, useMemo, useRef, useState } from 'react';
+import { useCallback, useMemo, useState } from 'react';
 import { useShallow } from 'zustand/react/shallow';
 import { useDesignerStore } from '@/features/bin-designer/store';
-import { useSettingsStore } from '@/core/store/settings';
-import { useToastStore } from '@/core/store/toast';
 import { DESIGNER_CONSTRAINTS, WALL_THICKNESS_OPTIONS } from '@/features/bin-designer/constants';
 import { Button, Stepper, InfoIcon } from '@/design-system';
 import { DeferredNumberInput } from '@/shared/components/DeferredNumberInput';
 import { SnappingSlider } from '../controls/SnappingSlider';
 import type { SnappingSliderOption } from '../controls/SnappingSlider';
 import {
-  getCompartmentCount,
-  getEligibleDividers,
-  isRectangularSelection,
-  cellIndex,
-  previewMergeCells,
-  previewSplitCells,
-} from '@/features/bin-designer/utils/compartments';
-import {
   minUniformCavity,
   solveCountForMinCavity,
 } from '@/features/bin-designer/utils/compartmentDimensions';
-import { getInteriorDims } from '@/features/bin-designer/utils/dividerAngle';
 import { useTranslation } from '@/i18n';
 import { useResponsive } from '@/shared/hooks/useResponsive';
-import { usePreviewColor } from '@/features/bin-designer/hooks/usePreviewColor';
-import { GridCell, GhostPreview } from './CompartmentEditorParts';
-import { DividerHitTargets } from './DividerHitTargets';
 import { DividerHeightControl } from './DividerHeightControl';
 import { DividerTiltSubsection } from './DividerTiltSubsection';
-import { rowKeyOf } from './useDividerTiltSubsection';
-import { useCompartmentLabeling } from './useCompartmentLabeling';
+import { CompartmentGridView } from './CompartmentGridView';
+import { useCompartmentGrid } from './useCompartmentGrid';
 import { getSegmentClass, SEGMENT_GROUP_CLASS } from '@/shared/components/segmentedControlClasses';
-
-const EMPTY_HIGHLIGHT_SET: ReadonlySet<number> = new Set();
 
 // Bounding box the 2D grid is scaled to fit, at the bin's true proportions.
 const GRID_ENVELOPE_W_PX = 360;
@@ -60,284 +47,38 @@ export function CompartmentEditor() {
   const t = useTranslation();
   const { isMobile } = useResponsive();
   const stepperSize = isMobile ? 'lg' : 'md';
+
+  const grid = useCompartmentGrid();
   const {
-    compartments,
-    labelSpan,
-    width,
-    depth,
-    gridUnitMm,
-    gridUnitMmY,
-    wallThickness,
-    style,
-    labelEnabled,
-    dividerTiltPreview,
-    selectedDividerKey,
-    hoveredDividerKey,
-    setParam,
-    setCompartmentGrid,
-    mergeCells,
-    splitCompartment,
-    updateLabel,
-    setPreviewCompartments,
-    setPreviewSelection,
-    setSelectedDividerKey,
-    setHoveredDividerKey,
-    setHoveredCompartmentId,
-  } = useDesignerStore(
+    cols,
+    rows,
+    thickness,
+    interiorW,
+    interiorD,
+    compartmentCount,
+    hasMergedCompartments,
+    aspectRatio,
+    labeling,
+    isDragging,
+    selectionAction,
+    hoveredIsSplittable,
+    instructionText,
+    applyGrid,
+    stepGrid,
+    handleThicknessChange,
+    handleReset,
+  } = grid;
+
+  const { labelSpan, labelEnabled, updateLabel } = useDesignerStore(
     useShallow((s) => ({
-      compartments: s.params.compartments,
       labelSpan: s.params.label.span === true,
-      width: s.params.width,
-      depth: s.params.depth,
-      gridUnitMm: s.params.gridUnitMm,
-      gridUnitMmY: s.params.gridUnitMmY,
-      wallThickness: s.params.wallThickness,
-      style: s.params.style,
       labelEnabled: s.params.label.enabled,
-      dividerTiltPreview: s.ui.dividerTiltPreview,
-      selectedDividerKey: s.ui.selectedDividerKey,
-      hoveredDividerKey: s.ui.hoveredDividerKey,
-      setParam: s.setParam,
-      setCompartmentGrid: s.setCompartmentGrid,
-      mergeCells: s.mergeCells,
-      splitCompartment: s.splitCompartment,
       updateLabel: s.updateLabel,
-      setPreviewCompartments: s.setPreviewCompartments,
-      setPreviewSelection: s.setPreviewSelection,
-      setSelectedDividerKey: s.setSelectedDividerKey,
-      setHoveredDividerKey: s.setHoveredDividerKey,
-      setHoveredCompartmentId: s.setHoveredCompartmentId,
     }))
   );
 
-  // Angled dividers are an advanced opt-in (see DividerTiltSubsection). Keep the
-  // on-grid hit-target overlay hidden until the user enables editing so dense
-  // grids stay legible (issue #2044).
-  const angledDividersEnabled = useSettingsStore((s) => s.settings.angledDividersEnabled);
-
-  const addToast = useToastStore((s) => s.addToast);
-
-  const { innerW: interiorW, innerD: interiorD } = getInteriorDims({
-    width,
-    depth,
-    gridUnitMm,
-    gridUnitMmY,
-    wallThickness,
-  });
-
-  const { cols, rows, thickness, cells } = compartments;
-
-  const compartmentCount = getCompartmentCount(compartments);
-
-  const labeling = useCompartmentLabeling(compartments, style, compartmentCount);
-
-  // Preview color synced with 3D preview (cross-tab + same-window CustomEvent)
-  const previewColor = usePreviewColor();
-
   // Manual size entry is an advanced opt-in; the grid steppers are the default.
   const [showSizer, setShowSizer] = useState(false);
-
-  // Selection state for drag-to-merge
-  const [selection, setSelection] = useState<Set<number>>(new Set());
-  const [dragStart, setDragStart] = useState<number | null>(null);
-  const [isDragging, setIsDragging] = useState(false);
-  const [hoverIdx, setHoverIdx] = useState<number | null>(null);
-  const gridRef = useRef<HTMLDivElement>(null);
-
-  // Mirror the hovered cell's compartment to the store so the 3D preview can
-  // draw that one compartment's dimension lines. Cleared while dragging (the
-  // ghost-merge preview takes over) and on unmount (cleanup nulls it).
-  useEffect(() => {
-    // Guard hoverIdx against the current grid: a resize can leave a stale
-    // hoverIdx pointing past the regenerated cells array.
-    const id = hoverIdx !== null && !isDragging && hoverIdx < cells.length ? cells[hoverIdx] : null;
-    setHoveredCompartmentId(id);
-    return () => {
-      setHoveredCompartmentId(null);
-    };
-  }, [hoverIdx, isDragging, cells, setHoveredCompartmentId]);
-
-  // Pre-compute cell counts per compartment to avoid repeated O(n) scans
-  const compartmentCellCounts = useMemo(() => {
-    const counts = new Map<number, number>();
-    for (const id of cells) {
-      counts.set(id, (counts.get(id) ?? 0) + 1);
-    }
-    return counts;
-  }, [cells]);
-
-  // Eligible divider segments for the hit-target overlay. Recomputed when the
-  // compartment grid changes (cheap O(rows*cols) scan, no perf concern).
-  const eligibleDividers = useMemo(() => getEligibleDividers(compartments), [compartments]);
-
-  // Compartments to brighten alongside the divider hover/selection — hover wins
-  // when both are set so power users can probe other dividers without losing
-  // their inspector context.
-  const dividerHighlightCompartments = useMemo<ReadonlySet<number>>(() => {
-    const activeKey = hoveredDividerKey ?? selectedDividerKey;
-    if (!activeKey) return EMPTY_HIGHLIGHT_SET;
-    const match = eligibleDividers.find(
-      (d) => rowKeyOf(d.compartmentA, d.compartmentB) === activeKey
-    );
-    return match ? new Set([match.compartmentA, match.compartmentB]) : EMPTY_HIGHLIGHT_SET;
-  }, [eligibleDividers, hoveredDividerKey, selectedDividerKey]);
-
-  // Stable label builder for hit-target aria-labels; uses the same wording as
-  // the panel's "edit row" affordance so screen readers hear a consistent name.
-  const rowLabelForHitTarget = useCallback(
-    (a: number, b: number) =>
-      t('binDesigner.angledDividers.editRowLabel', { a: String(a), b: String(b) }),
-    [t]
-  );
-
-  // Compute selection rectangle from drag start to current cell
-  const computeRectSelection = useCallback(
-    (startIdx: number, endIdx: number): Set<number> => {
-      const startCol = startIdx % cols;
-      const startRow = Math.floor(startIdx / cols);
-      const endCol = endIdx % cols;
-      const endRow = Math.floor(endIdx / cols);
-
-      const minCol = Math.min(startCol, endCol);
-      const maxCol = Math.max(startCol, endCol);
-      const minRow = Math.min(startRow, endRow);
-      const maxRow = Math.max(startRow, endRow);
-
-      const selected = new Set<number>();
-      for (let r = minRow; r <= maxRow; r++) {
-        for (let c = minCol; c <= maxCol; c++) {
-          selected.add(cellIndex(cols, c, r));
-        }
-      }
-      return selected;
-    },
-    [cols]
-  );
-
-  // Determine what action the current selection will trigger
-  const selectionAction = useMemo((): 'merge' | 'split' | 'none' => {
-    if (selection.size < 2) return 'none';
-    const indices = [...selection];
-    if (!isRectangularSelection(cols, indices)) return 'none';
-    const selectedIds = new Set(indices.map((i) => cells[i]));
-    return selectedIds.size === 1 ? 'split' : 'merge';
-  }, [selection, cols, cells]);
-
-  // Update preview compartments in store during drag for 3D ghost preview
-  const previewData = useMemo(() => {
-    if (!isDragging || selection.size < 2 || selectionAction === 'none') return null;
-
-    const indices = [...selection];
-
-    // Compute selection bounds
-    let minCol = cols,
-      maxCol = 0,
-      minRow = rows,
-      maxRow = 0;
-    for (const idx of indices) {
-      const col = idx % cols;
-      const row = Math.floor(idx / cols);
-      minCol = Math.min(minCol, col);
-      maxCol = Math.max(maxCol, col);
-      minRow = Math.min(minRow, row);
-      maxRow = Math.max(maxRow, row);
-    }
-
-    const newCells =
-      selectionAction === 'merge'
-        ? previewMergeCells(cells, indices)
-        : previewSplitCells(cells, indices);
-
-    return {
-      compartments: { cols, rows, thickness, cells: newCells },
-      selection: { action: selectionAction, minCol, maxCol, minRow, maxRow },
-    };
-  }, [isDragging, selection, selectionAction, cells, cols, rows, thickness]);
-
-  // Sync preview to store
-  useEffect(() => {
-    setPreviewCompartments(previewData?.compartments ?? null);
-    setPreviewSelection(previewData?.selection ?? null);
-    return () => {
-      setPreviewCompartments(null);
-      setPreviewSelection(null);
-    };
-  }, [previewData, setPreviewCompartments, setPreviewSelection]);
-
-  const handleCellPointerDown = useCallback(
-    (idx: number) => {
-      // Label mode: click selects for text entry. Setting no drag state keeps
-      // the merge/split pointer-up branch inert.
-      if (labeling.labelMode) {
-        labeling.selectCompartment(cells[idx]);
-        return;
-      }
-      setDragStart(idx);
-      setIsDragging(true);
-      setSelection(new Set([idx]));
-    },
-    [labeling, cells]
-  );
-
-  const handleCellPointerEnter = useCallback(
-    (idx: number) => {
-      setHoverIdx(idx);
-      if (!isDragging || dragStart === null) return;
-      setSelection(computeRectSelection(dragStart, idx));
-    },
-    [isDragging, dragStart, computeRectSelection]
-  );
-
-  const handleCellPointerLeave = useCallback(() => {
-    setHoverIdx(null);
-  }, []);
-
-  const handlePointerUp = useCallback(() => {
-    if (!isDragging) return;
-    setIsDragging(false);
-    setDragStart(null);
-
-    if (selection.size >= 2) {
-      const indices = [...selection];
-      if (isRectangularSelection(cols, indices)) {
-        // Check if all selected cells already belong to the same compartment
-        const selectedIds = new Set(indices.map((i) => cells[i]));
-        if (selectedIds.size === 1) {
-          // All same compartment — split it instead
-          splitCompartment(cells[indices[0]]);
-        } else {
-          // Different compartments — merge them
-          mergeCells(indices);
-        }
-      }
-    } else if (selection.size === 1) {
-      // Single cell click: if it's part of a multi-cell compartment, split it
-      const idx = [...selection][0];
-      const compartmentId = cells[idx];
-      if ((compartmentCellCounts.get(compartmentId) ?? 0) > 1) {
-        splitCompartment(compartmentId);
-      }
-    }
-
-    setSelection(new Set());
-  }, [isDragging, selection, cols, cells, mergeCells, splitCompartment, compartmentCellCounts]);
-
-  // All grid-dimension changes funnel through here so the "N labels cleared"
-  // warning (#2337) is shown consistently and the selection is reset once.
-  const applyGrid = useCallback(
-    (newCols: number, newRows: number) => {
-      const droppedLabels = setCompartmentGrid(newCols, newRows);
-      if (droppedLabels > 0) {
-        addToast({
-          message: t('binDesigner.compartmentEditor.labelsCleared', { count: droppedLabels }),
-          type: 'info',
-          duration: 4000,
-        });
-      }
-      setSelection(new Set());
-    },
-    [setCompartmentGrid, addToast, t]
-  );
 
   const handleColsChange = useCallback(
     (newCols: number) => {
@@ -346,17 +87,7 @@ export function CompartmentEditor() {
     [rows, applyGrid]
   );
 
-  const handleColsStep = useCallback(
-    (delta: number) => {
-      const next = cols + delta;
-      const clamped = Math.min(
-        DESIGNER_CONSTRAINTS.MAX_COMPARTMENT_GRID,
-        Math.max(DESIGNER_CONSTRAINTS.MIN_COMPARTMENT_GRID, next)
-      );
-      applyGrid(clamped, rows);
-    },
-    [cols, rows, applyGrid]
-  );
+  const handleColsStep = useCallback((delta: number) => stepGrid('cols', delta), [stepGrid]);
 
   const handleRowsChange = useCallback(
     (newRows: number) => {
@@ -365,17 +96,7 @@ export function CompartmentEditor() {
     [cols, applyGrid]
   );
 
-  const handleRowsStep = useCallback(
-    (delta: number) => {
-      const next = rows + delta;
-      const clamped = Math.min(
-        DESIGNER_CONSTRAINTS.MAX_COMPARTMENT_GRID,
-        Math.max(DESIGNER_CONSTRAINTS.MIN_COMPARTMENT_GRID, next)
-      );
-      applyGrid(cols, clamped);
-    },
-    [cols, rows, applyGrid]
-  );
+  const handleRowsStep = useCallback((delta: number) => stepGrid('rows', delta), [stepGrid]);
 
   // Size-led entry (fit-guarantee): typing a minimum opening picks the largest
   // count whose tightest compartment stays >= the requested mm, so every
@@ -424,19 +145,6 @@ export function CompartmentEditor() {
     [interiorD, thickness, cols, applyGrid]
   );
 
-  const handleThicknessChange = useCallback(
-    (newThickness: number) => {
-      setParam('compartments', { ...compartments, thickness: newThickness });
-    },
-    [compartments, setParam]
-  );
-
-  const handleReset = useCallback(() => {
-    applyGrid(cols, rows);
-  }, [cols, rows, applyGrid]);
-
-  const hasMergedCompartments = compartmentCount < cols * rows;
-
   // The mm fields show the smallest (worst-case interior) opening per axis — the
   // value the fit-guarantee is measured against. Rounded to 0.1mm for display so
   // the field never shows a stale or contradictory number relative to the grid.
@@ -456,35 +164,10 @@ export function CompartmentEditor() {
       ? 'input w-full h-12 text-center font-semibold tabular-nums'
       : 'w-full h-8 rounded border border-stroke-subtle bg-surface text-center text-sm tabular-nums text-content-secondary';
 
-  // Check if hovered cell is in a multi-cell compartment (splittable)
-  const hoveredIsSplittable = useMemo(() => {
-    if (hoverIdx === null || isDragging) return false;
-    const cId = cells[hoverIdx];
-    return (compartmentCellCounts.get(cId) ?? 0) > 1;
-  }, [hoverIdx, cells, isDragging, compartmentCellCounts]);
-
-  // Dynamic instruction text
-  const instructionText = useMemo(() => {
-    if (labeling.labelMode) {
-      return t('binDesigner.compartmentEditor.clickToLabel');
-    }
-    if (isDragging && selection.size >= 2) {
-      if (selectionAction === 'merge')
-        return t('binDesigner.compartmentEditor.releaseToMerge', { count: selection.size });
-      if (selectionAction === 'split') return t('binDesigner.compartmentEditor.releaseToSplit');
-      return t('binDesigner.compartmentEditor.dragToSelect');
-    }
-    if (hoveredIsSplittable && !isDragging) {
-      return t('binDesigner.compartmentEditor.clickToSplit');
-    }
-    return t('binDesigner.compartmentEditor.dragOrClick');
-  }, [labeling.labelMode, isDragging, selection.size, selectionAction, hoveredIsSplittable, t]);
-
   // Render the grid at the bin's true top-view proportions, scaled to fit a
   // fixed envelope. Capping width to `MAX_H * aspect` keeps the derived height
   // within budget while preserving the real shape — so deep bins stay legible
   // instead of ballooning, and the box never overflows onto the controls above.
-  const aspectRatio = depth > 0 ? width / depth : 1;
   const gridMaxWidthPx = Math.min(GRID_ENVELOPE_W_PX, GRID_ENVELOPE_H_PX * aspectRatio);
 
   // Build wall thickness options for SnappingSlider
@@ -496,17 +179,6 @@ export function CompartmentEditor() {
       })),
     [t]
   );
-
-  // Compute grid intersection dots (internal intersections only)
-  const gridDots = useMemo(() => {
-    const dots: Array<{ x: number; y: number }> = [];
-    for (let row = 1; row < rows; row++) {
-      for (let col = 1; col < cols; col++) {
-        dots.push({ x: col / cols, y: row / rows });
-      }
-    }
-    return dots;
-  }, [cols, rows]);
 
   return (
     <div className="space-y-5">
@@ -697,106 +369,13 @@ export function CompartmentEditor() {
               </Button>
             )}
           </div>
-          <div
-            ref={gridRef}
-            className="relative mr-auto w-full select-none overflow-hidden rounded-lg border-2 border-stroke-subtle bg-surface-elevated p-2"
-            style={{ aspectRatio, maxWidth: `${gridMaxWidthPx}px` }}
-            role="application"
-            aria-label={`Compartment grid, ${cols} columns by ${rows} rows`}
-            aria-describedby="compartment-grid-instructions"
-            onPointerUp={handlePointerUp}
-            onPointerLeave={() => {
-              handlePointerUp();
-              setHoverIdx(null);
-            }}
-          >
-            {/* Grid dots overlay at intersections */}
-            {gridDots.length > 0 && (
-              <div className="pointer-events-none absolute inset-2">
-                {gridDots.map(({ x, y }, i) => (
-                  <div
-                    key={i}
-                    className="absolute h-1.5 w-1.5 -translate-x-1/2 -translate-y-1/2 rounded-full bg-content-tertiary/40"
-                    style={{ left: `${x * 100}%`, top: `${(1 - y) * 100}%` }}
-                  />
-                ))}
-              </div>
-            )}
-            {/* Use flex-col-reverse to match 3D orientation: row 0 = front = bottom of UI */}
-            {/* No gap - merged cells should visually connect; borders handle separation */}
-            <div className="relative flex h-full w-full flex-col-reverse">
-              {Array.from({ length: rows }, (_, visualRow) => {
-                const dataRow = visualRow; // flex-col-reverse handles the flip
-                return (
-                  <div
-                    key={dataRow}
-                    className="grid min-h-0 min-w-0 flex-1"
-                    style={{
-                      gridTemplateColumns: `repeat(${cols}, 1fr)`,
-                    }}
-                  >
-                    {Array.from({ length: cols }, (_, col) => {
-                      const idx = dataRow * cols + col;
-                      const compartmentId = cells[idx];
-                      return (
-                        <GridCell
-                          key={idx}
-                          idx={idx}
-                          compartmentId={compartmentId}
-                          isSelected={selection.has(idx)}
-                          isHovered={hoverIdx === idx && !isDragging}
-                          isSplittable={
-                            !isDragging && (compartmentCellCounts.get(compartmentId) ?? 0) > 1
-                          }
-                          isDragging={isDragging}
-                          isDividerHoverHighlighted={dividerHighlightCompartments.has(
-                            compartmentId
-                          )}
-                          labelMode={labeling.labelMode}
-                          isLabelSelected={
-                            labeling.labelMode && labeling.editingId === compartmentId
-                          }
-                          labelText={labeling.textOf(compartmentId)}
-                          displayNumber={labeling.displayNumberOf(compartmentId)}
-                          config={compartments}
-                          previewColor={previewColor}
-                          onPointerDown={handleCellPointerDown}
-                          onPointerEnter={handleCellPointerEnter}
-                          onPointerLeave={handleCellPointerLeave}
-                        />
-                      );
-                    })}
-                  </div>
-                );
-              })}
-            </div>
-            {/* Divider hit targets: clickable lines above the cells, transparent
-                container with per-line pointer-events so cell drag-merge still
-                works. Hidden during cell-merge drag to keep the surface calm. */}
-            {!isDragging && angledDividersEnabled && eligibleDividers.length > 0 && (
-              <DividerHitTargets
-                compartments={compartments}
-                dividers={eligibleDividers}
-                interiorW={interiorW}
-                interiorD={interiorD}
-                preview={dividerTiltPreview}
-                selectedKey={selectedDividerKey}
-                hoveredKey={hoveredDividerKey}
-                onSelect={setSelectedDividerKey}
-                onHoverChange={setHoveredDividerKey}
-                rowLabel={rowLabelForHitTarget}
-              />
-            )}
-            {/* Ghost preview overlay during cell-select drag */}
-            {isDragging && selection.size >= 2 && (
-              <GhostPreview
-                selection={selection}
-                selectionAction={selectionAction}
-                cols={cols}
-                rows={rows}
-              />
-            )}
-          </div>
+
+          <CompartmentGridView
+            grid={grid}
+            className="mr-auto w-full"
+            style={{ maxWidth: `${gridMaxWidthPx}px` }}
+            describedById="compartment-grid-instructions"
+          />
 
           {labeling.labelMode && (
             <>
